% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{cooler_resolutions}
\alias{cooler_resolutions}
\title{cooler_resolutions}
\usage{
cooler_resolutions(mcool)
}
\arguments{
\item{file}{path to .cool or m.cool file}
}
\value{
list of resolutions in cooler
}
\description{
get resolutions from a mcool file
}
\author{
Marcin Imielinski, Stanley Clarke
}
\keyword{cooler_resolutions}
