% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{getBins}
\alias{getBins}
\title{getBins}
\usage{
getBins(file, res = NULL)
}
\arguments{
\item{file}{path to .cool or m.cool file}

\item{res}{resolution to use that is in the mcool file}
}
\description{
get the bins of the mcool file for a specific mcool file at a given resolution
}
\author{
Stanley Clarke
}
\keyword{getBins}
