% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GxG.R
\name{GxG}
\alias{GxG}
\alias{gMatrix}
\title{GxG}
\description{
Matrices and pairwise features in a GenomicRanges / data.table framework

Copyright (C) 2018 Marcin Imielinski

   This program is free software: you can redistribute it and/or modify
   it under the terms of the GNU General Public License as published by
   the Free Software Foundation, either version 3 of the License, or
   (at your option) any later version.

   This program is distributed in the hope that it will be useful,
   but WITHOUT ANY WARRANTY; without even the implied warranty of
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
   GNU General Public License for more details.

   You should have received a copy of the GNU General Public License
   along with this program.  If not, see <http://www.gnu.org/licenses/>.

   Github: https://github.com/mskilab/GxG
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gMatrix-new}{\code{gMatrix$new()}}
\item \href{#method-gMatrix-disjoin}{\code{gMatrix$disjoin()}}
\item \href{#method-gMatrix-sweep}{\code{gMatrix$sweep()}}
\item \href{#method-gMatrix-metaset}{\code{gMatrix$metaset()}}
\item \href{#method-gMatrix-annotate}{\code{gMatrix$annotate()}}
\item \href{#method-gMatrix-clusters}{\code{gMatrix$clusters()}}
\item \href{#method-gMatrix-transform}{\code{gMatrix$transform()}}
\item \href{#method-gMatrix-set}{\code{gMatrix$set()}}
\item \href{#method-gMatrix-agg}{\code{gMatrix$agg()}}
\item \href{#method-gMatrix-merge}{\code{gMatrix$merge()}}
\item \href{#method-gMatrix-subset}{\code{gMatrix$subset()}}
\item \href{#method-gMatrix-gtrack}{\code{gMatrix$gtrack()}}
\item \href{#method-gMatrix-peaks}{\code{gMatrix$peaks()}}
\item \href{#method-gMatrix-hicpro}{\code{gMatrix$hicpro()}}
\item \href{#method-gMatrix-pack}{\code{gMatrix$pack()}}
\item \href{#method-gMatrix-print}{\code{gMatrix$print()}}
\item \href{#method-gMatrix-clone}{\code{gMatrix$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-new"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$new(
  gr = NULL,
  dat = NULL,
  field = NULL,
  fill = 0,
  full = NULL,
  na.rm = TRUE,
  agg.fun = sum,
  lower.tri = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-disjoin"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-disjoin}{}}}
\subsection{Method \code{disjoin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$disjoin(gr = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-sweep"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-sweep}{}}}
\subsection{Method \code{sweep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$sweep(vec, FUN = "-", MARGIN = c(1))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-metaset"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-metaset}{}}}
\subsection{Method \code{metaset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$metaset(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-annotate"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-annotate}{}}}
\subsection{Method \code{annotate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$annotate(y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-clusters"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-clusters}{}}}
\subsection{Method \code{clusters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$clusters(mode = "louvain", gpair = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-transform"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$transform(FUN, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-set"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$set(val, gr1 = NULL, gr2 = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-agg"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-agg}{}}}
\subsection{Method \code{agg()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$agg(gr = NULL, FUN = private$pagg.fun, weighted = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-merge"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-merge}{}}}
\subsection{Method \code{merge()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$merge(
  gmats = NULL,
  FUN = self$agg.fun,
  simplify = FALSE,
  mc.cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-subset"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-subset}{}}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$subset(gr1 = NULL, gr2 = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-gtrack"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-gtrack}{}}}
\subsection{Method \code{gtrack()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$gtrack(
  name = "",
  colormap = c("white", "red", "black"),
  clim = NA,
  cmap.min = NA,
  cmap.max = NA,
  quantile = 0.01,
  max.ranges = 3000,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-peaks"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-peaks}{}}}
\subsection{Method \code{peaks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$peaks(reverse = FALSE, full = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-hicpro"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-hicpro}{}}}
\subsection{Method \code{hicpro()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$hicpro(file, verbose = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-pack"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-pack}{}}}
\subsection{Method \code{pack()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$pack(seqname = "1")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-print"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gMatrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gMatrix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gMatrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
